Move record-normalizer to its own module
