+++ packages/@orbit/records/src/record-normalizer.ts
@@ -1,0 +1,13 @@
+import {
+  InitializedRecord,
+  RecordIdentity,
+  UninitializedRecord
+} from './record';
+
+export interface RecordNormalizer<
+  RI = RecordIdentity,
+  R = UninitializedRecord
+> {
+  normalizeRecordIdentity(recordIdentity: RI): RecordIdentity;
+  normalizeRecord(record: R): InitializedRecord;
+}
