+++ packages/@orbit/records/src/record.ts
@@ -61,14 +61,6 @@ export interface RecordInitializer {
   initializeRecord(record: UninitializedRecord): InitializedRecord;
 }
 
-export interface RecordNormalizer<
-  RI = RecordIdentity,
-  R = UninitializedRecord
-> {
-  normalizeRecordIdentity(recordIdentity: RI): RecordIdentity;
-  normalizeRecord(record: R): InitializedRecord;
-}
-
 export function cloneRecordIdentity(identity: RecordIdentity): RecordIdentity {
   const { type, id } = identity;
   return { type, id };
